refactor: code quality improvements and documentation standardisation
This commit addresses code quality enhancements, British English
standardisation, and documentation accuracy improvements across the
codebase.

## Code Quality Improvements

### command_mode.c
- Extracted apply_log_level() helper function to eliminate code duplication
  in log level selection (lines 692-713)
- Added comprehensive docstring for apply_log_level() helper
- Improved timeout handling logic with clearer state-based messaging
- British English: color → colour throughout comments

### config_storage.c/h
- Fixed v1 version size calculation in get_config_size_for_version()
  - v1 included log_level only (not led_brightness)
  - Changed return from offsetof(keyboard_id) to offsetof(led_brightness)
  - Corrects migration behaviour for v1→v2 upgrades
- Moved led_brightness from "v1+" to "v2+" section in structure comments
- British English: behavior → behaviour in comments

### hid_interface.c
- Added explicit uint8_t casts in handle_mouse_report() button bit-shifting
  to prevent implicit integer promotion warnings from static analyzers
- Added comment explaining cast rationale

## Documentation Improvements

### config-storage.md
- Corrected Pico SDK flash operation description (lines 304-310)
  - Clarified SDK places functions in RAM at build time (linker attributes)
  - Listed application responsibilities: disable interrupts, other core not
    executing from flash, source data in RAM
  - Mentioned flash_safe_execute() helper and this firmware's compliance
- Updated version history to correctly reflect v1 (log_level only) vs v2
  (added led_brightness, keyboard_id, shift_override_enabled)
- Removed unverified quantification "frequently"

### British English Standardisation
Applied British spelling throughout 60 files:
- color → colour
- behavior → behaviour
- Consistent terminology in comments and documentation

## Configuration Files

### .clang-format (new file)
Added clang-format configuration matching project standards:
- BasedOnStyle: Google
- IndentWidth: 4
- ColumnLimit: 100
- AlignConsecutive*: true (for embedded systems readability)

## Miscellaneous

- Multiple documentation files: removed marketing language, improved
  technical accuracy, British English consistency
- Comment improvements across protocol handlers, LED helpers, UART driver
- No functional changes to runtime behaviour

+++ src/common/lib/hid_interface.c
@@ -486,8 +486,10 @@ void send_empty_keyboard_report(void) {
  */
 void handle_mouse_report(const uint8_t buttons[5], int8_t pos[3]) {
     // Handle Mouse Report
+    // Cast each shifted uint8_t explicitly to avoid implicit integer promotion warnings
     mouse_report.buttons =
-        buttons[0] | (buttons[1] << 1) | (buttons[2] << 2) | (buttons[3] << 3) | (buttons[4] << 4);
+        (uint8_t)(buttons[0] | (uint8_t)(buttons[1] << 1) | (uint8_t)(buttons[2] << 2) |
+                  (uint8_t)(buttons[3] << 3) | (uint8_t)(buttons[4] << 4));
     mouse_report.x     = pos[0];
     mouse_report.y     = pos[1];
     mouse_report.wheel = pos[2];
